fix(test): don't remove .gitkeep from test user presets dir

+++ EdenSynth/SharedCode_test/source/presets_test/PresetsTest.cpp
@@ -365,7 +365,9 @@ TEST(FileUserPresetsDataSource, SavesAndLoadsPresetsToDisk) {
   const wolfsound::WhenLeavingScopeExecute cleanup{[&] {
     std::for_each(std::filesystem::directory_iterator{userPresetsPath},
                   std::filesystem::directory_iterator{}, [](const auto& entry) {
-                    std::filesystem::remove(entry.path());
+                    if (entry.path().extension() == ".json") {
+                      std::filesystem::remove(entry.path());
+                    }
                   });
   }};
